+++ src/searches/benchmarks.hpp
@@ -1,14 +1,13 @@
 #ifndef REMAINDERTREE_SRC_SEARCHES_BENCHMARKS_H_
 #define REMAINDERTREE_SRC_SEARCHES_BENCHMARKS_H_
 
-//std::function< std::vector<bool> (int N)> remember this is what is being passed in!
+//std::function< std::vector<bool> (long N)> remember this is what is being passed in!
 #include <vector>
 
 
-std::vector<bool> constant_slow(int);
+std::vector<bool> constant_slow(long);
 
-std::vector<bool> random_zz(int, int, int);
+std::vector<bool> random_zz(long, long, long);
 
-#include "benchmarks.cpp" //TODO: this is horrible form! not necessary as before because not templates
 
 #endif //REMAINDERTREE_SRC_SEARCHES_BENCHMARKS_H_
